+++ CHANGES.rst
@@ -3,7 +3,8 @@ Changelog of lizard-nxt client
 
 Unreleased (2.0.3) (XXXX-XX-XX)
 ---------------------
--
+
+- Allow asset layers to have different name than 'waterchain', as long as 'waterchain' is in the layergroup slug, it should work. Grid layers should be named <layegroup_slug>_grid.
 
 
 Release 2.0.8 (2015-9-4)